requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.22.0
python-multipart
PyJWT
python-dotenv
mariadb
requests

